% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhtdensity.R
\docType{package}
\name{dhtdensity}
\alias{dhtdensity}
\alias{dhtdensity-package}
\title{dhtdensity: model densities with Bayesian hierarchical models}
\description{
This package models densities using MCMC.
}
