% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-predictive.R
\name{post_pred}
\alias{post_pred}
\title{Sample from the posterior predictive distribution}
\usage{
post_pred(z_grid, n_rep, n_warmup, mcmc_output, time)
}
\arguments{
\item{z_grid}{A grid of depths to evaluate the function over.}

\item{n_rep}{Number of posterior predictive samples.}

\item{n_warmup}{Number of warmup iterations. Must not be on the global scale,
but on the actual number of total iterations scale.}

\item{mcmc_output}{A list output from \code{sample_dht_mcmc} or
\code{sample_dht_mcmc_ind}.}

\item{time}{A time to compute the posterior predictive distribution for.}
}
\value{
A matrix of posterior predictive draws. Rows are individual density curves.
}
\description{
Sample from the posterior predictive distribution
}
