% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Read the raw data into R}
\usage{
read_data()
}
\value{
A named list of the density measurements \code{y}, and the depths \code{z}.
}
\description{
Read the raw data into R
}
