% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sampler.R
\name{sample_dht_mcmc}
\alias{sample_dht_mcmc}
\title{MCMC sampler for the DHT hierarchical model}
\usage{
sample_dht_mcmc(y, z, n_sample, chol_prop_mat, inits = list(alpha = NULL,
  beta = NULL, mu_alpha = NULL, mu_beta = NULL, sigma2_alpha = NULL,
  sigma2_beta = NULL, sigma2_y = NULL), priors = list(mu_alpha_mean =
  NULL, mu_beta_mean = NULL, mu_alpha_sd = NULL, mu_beta_sd = NULL,
  sigma2_max = 5000), control = list(thin = 1, progress = FALSE))
}
\arguments{
\item{y}{Density data. Must be a matrix with rows corresponding to individual
groups.}

\item{z}{Grid of depths. Must be a vector of length the same as \code{ncol(y)}.}

\item{n_sample}{Number of MCMC samples.}

\item{chol_prop_mat}{An array of cholesky's of the desired proposal matrix for the
MH step for the beta's (nonlinear dht parameters).}

\item{inits}{A list of initial values. Can contain \code{alpha}, \code{beta},
\code{mu_alpha}, \code{mu_beta}, \code{sigma2_alpha}, \code{sigma2_beta}, 
and \code{sigma2_y}. Any values which are unspecified will take the
default values.}

\item{priors}{A list of prior values. Can contain \code{mu_alpha_mean}, \code{mu_alpha_mean},
\code{mu_beta_sd}, \code{mu_beta_sd}, and \code{sigma2_max}. Any values which
are unspecified will take the defaults.}

\item{control}{A list of control values. Can contain \code{thin}, and \code{progress}.}
}
\value{
A named list containing elements \code{alpha}, \code{beta},
\code{mu_alpha}, \code{mu_beta}, \code{sigma2_alpha}, \code{sigma2_beta},
and \code{sigma2_y}.
}
\description{
MCMC sampler for the DHT hierarchical model
}
