% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-sampler-no-hierarchy.R
\name{sample_dht_mcmc_ind}
\alias{sample_dht_mcmc_ind}
\title{MCMC sampler for non-hierarchical model (no hierarchy)}
\usage{
sample_dht_mcmc_ind(y, z, n_sample, n_adapt, n_adapt_interval, n_warmup,
  chol_prop_mat, priors = list(alpha_mean = NULL, alpha_sd = NULL,
  beta_mean = NULL, beta_sd = NULL), inits = list(alpha = NULL, beta =
  NULL, sigma2_y = NULL), control = list(thin = 1, progress = FALSE))
}
\arguments{
\item{y}{Vector of observed values}

\item{n_sample}{Number of MCMC samples.}

\item{n_adapt}{Number of adaptation samples}

\item{n_adapt_interval}{Interval between adaptations.}

\item{n_warmup}{Number of warmup iterations.}

\item{priors}{List of prior values. Can contain \code{beta_prior_mean} and
\code{beta_prior_sd}.}

\item{inits}{List of initial values. Can contain \code{alpha}, \code{beta},
and \code{sigma2_y}. If any unspecified reasonable random starting values 
will be generated.}
}
\value{
A list of all MCMC samples of alpha, beta, and sigma2_y.
}
\description{
MCMC sampler for non-hierarchical model (no hierarchy)
}
