% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_double_tanh}
\alias{simulate_double_tanh}
\title{Simulate data according to the dht model}
\usage{
simulate_double_tanh(z, n_groups, sigma2_y, mu_alpha, sigma2_alpha,
  mu_beta, sigma2_beta)
}
\arguments{
\item{z}{Depths at which to evaluate the density.}

\item{n_groups}{Number of groups to simulate.}

\item{sigma2_y}{Variance about the mean curve.}

\item{mu_alpha}{Mean value of the alpha's.}

\item{sigma2_alpha}{Variance of the alpha's.}

\item{mu_beta}{Mean value of the beta's.}

\item{sigma2_beta}{Variance of the beta's.}
}
\value{
A matrix of simulated data. Rows are the individual density curves.
}
\description{
Simulate data according to the dht model
}
